ous_agile_devops_ai/oad_ai/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
Django>=5.0.6
django-cors-headers==4.4.0
django-rest-auth==0.9.5
django-saml2-auth==2.2.1
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangosaml2==1.9.3
elementpath==4.4.0
idna==3.7
lxml==5.2.2
path==17.0.0
pip==23.2.1
pycparser==2.22
PyJWT==1.7.1
pyOpenSSL==24.2.1
pysaml2==7.5.0
channels==4.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
setuptools==68.2.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
views==0.3
wheel==0.41.2
whitenoise==6.7.0
xmlschema==2.5.1
xmlsec==1.3.14
djangorestframework-simplejwt==5.2.2
djangosaml2==1.9.3
django-csp==3.7
schedule==1.2.0
